+++ HrothCore/src/HrothCore/Core/IClient.hpp
@@ -9,7 +9,7 @@ namespace HrothCore
 
             virtual void Init() = 0;
             virtual void Shutdown() = 0;
-            virtual void Update(double dt) = 0;
+            virtual void Update(float dt) = 0;
 
             virtual void ImGuiRender() = 0;
     };
